Reorganizaed fake provider
