+++ tests/Feature/Dispatcher/ContactDispatcherTest.php
@@ -3,9 +3,6 @@
 namespace TheTreehouse\Relay\Tests\Feature\Dispatcher;
 
 use TheTreehouse\Relay\Tests\Fixtures\Models\Contact;
-use TheTreehouse\Relay\Tests\Fixtures\Providers\FakeProvider\Jobs\CreateFakeContact;
-use TheTreehouse\Relay\Tests\Fixtures\Providers\FakeProvider\Jobs\DeleteFakeContact;
-use TheTreehouse\Relay\Tests\Fixtures\Providers\FakeProvider\Jobs\UpdateFakeContact;
 
 /** @group feature.dispatcher */
 class ContactDispatcherTest extends BaseDispatcherTest
@@ -13,10 +10,4 @@ class ContactDispatcherTest extends BaseDispatcherTest
     protected $entityName = 'Contact';
 
     protected $entityModelClass = Contact::class;
-
-    protected $createJob = CreateFakeContact::class;
-
-    protected $updateJob = UpdateFakeContact::class;
-
-    protected $deleteJob = DeleteFakeContact::class;
 }
